Moved Input/Output/Tensor descriptors to ov (#7219)
* Moved Input/Output/Tensor descriptors to ov

* Fixed build
