+++ ngraph/core/src/descriptor/tensor.cpp
@@ -2,61 +2,60 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "ngraph/descriptor/tensor.hpp"
+#include "openvino/core/descriptor/tensor.hpp"
 
 #include "ngraph/node.hpp"
 
-using namespace ngraph;
 using namespace std;
 
-descriptor::Tensor::Tensor(const element::Type& element_type, const PartialShape& pshape, const std::string& name)
+ov::descriptor::Tensor::Tensor(const element::Type& element_type, const PartialShape& pshape, const std::string& name)
     : m_element_type(element_type),
       m_partial_shape(pshape),
       m_name(name),
       m_shape_changed(true) {}
 
-descriptor::Tensor::Tensor(const element::Type& element_type,
-                           const PartialShape& pshape,
-                           Node* node,
-                           size_t node_output_number)
+ov::descriptor::Tensor::Tensor(const element::Type& element_type,
+                               const PartialShape& pshape,
+                               ngraph::Node* node,
+                               size_t node_output_number)
     : m_element_type(element_type),
       m_partial_shape(pshape),
       m_shape_changed(true) {}
 
-void descriptor::Tensor::set_tensor_type(const element::Type& element_type, const PartialShape& pshape) {
+void ov::descriptor::Tensor::set_tensor_type(const element::Type& element_type, const PartialShape& pshape) {
     set_element_type(element_type);
     set_partial_shape(pshape);
 }
 
-void descriptor::Tensor::set_element_type(const element::Type& element_type) {
+void ov::descriptor::Tensor::set_element_type(const element::Type& element_type) {
     m_element_type = element_type;
 }
 
-void descriptor::Tensor::set_partial_shape(const PartialShape& partial_shape) {
+void ov::descriptor::Tensor::set_partial_shape(const PartialShape& partial_shape) {
     m_partial_shape = partial_shape;
     m_shape_changed = true;
 }
 
-void descriptor::Tensor::invalidate_values() {
+void ov::descriptor::Tensor::invalidate_values() {
     m_upper_value = nullptr;
     m_lower_value = nullptr;
 }
 
-void descriptor::Tensor::set_lower_value(const HostTensorPtr& value) {
+void ov::descriptor::Tensor::set_lower_value(const ngraph::HostTensorPtr& value) {
     NGRAPH_CHECK(value != nullptr);
     NGRAPH_CHECK(m_partial_shape.same_scheme(value->get_partial_shape()));
     NGRAPH_CHECK(m_element_type == value->get_element_type());
     m_lower_value = value;
 }
 
-void descriptor::Tensor::set_upper_value(const HostTensorPtr& value) {
+void ov::descriptor::Tensor::set_upper_value(const ngraph::HostTensorPtr& value) {
     NGRAPH_CHECK(value != nullptr);
     NGRAPH_CHECK(m_partial_shape.same_scheme(value->get_partial_shape()));
     NGRAPH_CHECK(m_element_type == value->get_element_type());
     m_upper_value = value;
 }
 
-const Shape& descriptor::Tensor::get_shape() const {
+const ngraph::Shape& ov::descriptor::Tensor::get_shape() const {
     if (m_partial_shape.is_static()) {
         if (m_shape_changed.load(std::memory_order_relaxed)) {
             std::lock_guard<std::mutex> guard(shape_mutex);
@@ -72,7 +71,7 @@ const Shape& descriptor::Tensor::get_shape() const {
     }
 }
 
-size_t descriptor::Tensor::size() const {
+size_t ov::descriptor::Tensor::size() const {
     const bool bitwidth_less_than_byte = m_element_type.bitwidth() < 8;
     if (bitwidth_less_than_byte) {
         return ceil((1.0 * shape_size(get_shape()) * m_element_type.bitwidth()) / 8);
@@ -81,30 +80,30 @@ size_t descriptor::Tensor::size() const {
 }
 
 NGRAPH_SUPPRESS_DEPRECATED_START
-void descriptor::Tensor::set_name(const string& name) {
+void ov::descriptor::Tensor::set_name(const string& name) {
     m_name = name;
 }
 
-const std::string& descriptor::Tensor::get_name() const {
+const std::string& ov::descriptor::Tensor::get_name() const {
     return m_name;
 }
 NGRAPH_SUPPRESS_DEPRECATED_END
 
-const std::unordered_set<std::string>& descriptor::Tensor::get_names() const {
+const std::unordered_set<std::string>& ov::descriptor::Tensor::get_names() const {
     return m_names;
 }
 
-void descriptor::Tensor::set_names(const std::unordered_set<std::string>& names) {
+void ov::descriptor::Tensor::set_names(const std::unordered_set<std::string>& names) {
     m_names = names;
 }
 
-void descriptor::Tensor::add_names(const std::unordered_set<std::string>& names) {
+void ov::descriptor::Tensor::add_names(const std::unordered_set<std::string>& names) {
     for (const auto& name : names) {
         m_names.insert(name);
     }
 }
 
-ostream& operator<<(ostream& out, const descriptor::Tensor& tensor) {
+ostream& operator<<(ostream& out, const ov::descriptor::Tensor& tensor) {
     std::string names;
     for (const auto& name : tensor.get_names()) {
         if (!names.empty())
